+++ src/OrbitQt/CMakeLists.txt
@@ -20,6 +20,7 @@ target_sources(
           FilterPanelWidgetAction.h
           HistogramWidget.h
           MultipleOfSpinBox.h
+          MultipleOfValidator.h
           orbitaboutdialog.h
           orbitdataviewpanel.h
           orbiteventiterator.h
@@ -50,6 +51,7 @@ target_sources(
           FilterPanelWidgetAction.cpp
           HistogramWidget.cpp
           MultipleOfSpinBox.cpp
+          MultipleOfValidator.cpp
           StatusListenerImpl.cpp
           orbitaboutdialog.cpp
           orbitaboutdialog.ui
@@ -118,7 +120,7 @@ target_sources(OrbitQtTests
                 AnnotatingSourceCodeDialogTest.cpp
                 CallTreeViewItemModelTest.cpp
                 HistogramWidgetTest.cpp
-        MultipleOfValidatorTest.cpp
+                MultipleOfValidatorTest.cpp
                 StatusListenerImplTest.cpp
                 TutorialOverlayTest.cpp
                 TrackTypeItemModelTest.cpp)
