+++ src/OrbitQt/MultipleOfSpinBox.h
@@ -1,8 +1,8 @@
 // Copyright (c) 2022 The Orbit Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-#ifndef ORBIT_QT_Q_MULTIPLE_OF_EIGHT_SPIN_BOX_H_
-#define ORBIT_QT_Q_MULTIPLE_OF_EIGHT_SPIN_BOX_H_
+#ifndef ORBIT_QT_Q_MULTIPLE_OF_SPIN_BOX_H_
+#define ORBIT_QT_Q_MULTIPLE_OF_SPIN_BOX_H_
 
 #include <QSpinBox>
 #include <memory>
@@ -28,15 +28,11 @@ class MultipleOfValidator : public QIntValidator {
 class MultipleOfSpinBox : public QSpinBox {
  public:
   explicit MultipleOfSpinBox(QWidget* parent = nullptr) : QSpinBox(parent) {}
-  void SetValidator(std::unique_ptr<QIntValidator> validator) { validator_ = std::move(validator); }
 
  protected:
   QValidator::State validate(QString& input, int& pos) const override;
-
- private:
-  std::unique_ptr<QIntValidator> validator_;
 };
 
 }  // namespace orbit_qt
 
-#endif  // ORBIT_QT_Q_MULTIPLE_OF_EIGHT_SPIN_BOX_H_
+#endif  // ORBIT_QT_Q_MULTIPLE_OF_SPIN_BOX_H_
