+++ src/OrbitQt/MultipleOfValidator.h
@@ -1,0 +1,25 @@
+#ifndef ORBIT_QT_Q_MULTIPLE_OF_VALIDATOR_H_
+#define ORBIT_QT_Q_MULTIPLE_OF_VALIDATOR_H_
+
+#include <QIntValidator>
+#include <QObject>
+
+namespace orbit_qt {
+
+// A specialization of `QIntValidator` that only accepts values that are a multiple of
+// `multiple_of_`.
+class MultipleOfValidator : public QIntValidator {
+ public:
+  explicit MultipleOfValidator(QObject* parent = nullptr) : QIntValidator(parent) {}
+  MultipleOfValidator(int bottom, int top, int multiple_of, QObject* parent = nullptr)
+      : QIntValidator(bottom, top, parent), multiple_of_{multiple_of} {}
+  void SetMultipleOf(int multiple_of) { multiple_of_ = multiple_of; }
+  [[nodiscard]] QValidator::State validate(QString& input, int& pos) const override;
+
+ private:
+  int multiple_of_ = 1;
+};
+
+}  // namespace orbit_qt
+
+#endif  // ORBIT_QT_Q_MULTIPLE_OF_VALIDATOR_H_
